update: register confirm password

+++ frontend/src/components/Register.jsx
@@ -1,33 +1,39 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 const Register = (props) => {
-  const [email, setEmail] = React.useState('');
-  const [password, setPassword] = React.useState('');
-  const [name, setName] = React.useState('');
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [name, setName] = useState('');
   const navigate = useNavigate();
 
+  // registering event
   const register = async () => {
-    const response = await fetch('http://localhost:5005/user/auth/register', {
-      method: 'POST',
-      body: JSON.stringify({
-        email,
-        password,
-        name
-      }),
-      headers: {
-        'Content-type': 'application/json',
-      }
-    });
-    const data = await response.json();
+    if (password === setConfirmPassword) {
+      const response = await fetch('http://localhost:5005/user/auth/register', {
+        method: 'POST',
+        body: JSON.stringify({
+          email,
+          password,
+          name
+        }),
+        headers: {
+          'Content-type': 'application/json',
+        }
+      });
+      const data = await response.json();
 
-    if (data.error) {
-      alert(data.error);
-    } else if (data.token) {
-      localStorage.setItem('token', data.token);
-      localStorage.setItem('email', email);
-      props.setToken(data.token);
-      navigate('/');
+      if (data.error) {
+        alert(data.error);
+      } else if (data.token) {
+        localStorage.setItem('token', data.token);
+        localStorage.setItem('email', email);
+        props.setToken(data.token);
+        navigate('/');
+      }
+    } else {
+      alert('Passwords not mathcing')
     }
   };
 
@@ -43,7 +49,9 @@ const Register = (props) => {
 
               <form className="space-y-4 md:space-y-6" action="#">
                 <div>
-                  <label htmlFor="userName" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Your name</label>
+                  <label htmlFor="userName" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
+                    Your name
+                  </label>
                   <input
                     type="text"
                     name="userName"
@@ -57,7 +65,9 @@ const Register = (props) => {
                 </div>
 
                 <div>
-                  <label htmlFor="email" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Your email</label>
+                  <label htmlFor="email" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
+                    Your email
+                  </label>
                   <input
                     type="email"
                     name="email"
@@ -71,7 +81,9 @@ const Register = (props) => {
                 </div>
 
                 <div>
-                  <label htmlFor="password" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Password</label>
+                  <label htmlFor="password" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
+                    Password
+                  </label>
                   <input
                     type="password"
                     name="password"
@@ -81,6 +93,22 @@ const Register = (props) => {
                     required
                     value={password}
                     onChange={e => setPassword(e.target.value)}
+                  />
+                </div>
+
+                <div>
+                  <label htmlFor="confirmPassword" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
+                    Confirm password
+                  </label>
+                  <input
+                    type="password"
+                    name="confirmPassword"
+                    id="confirmPassword"
+                    placeholder="••••••••"
+                    className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                    required
+                    value={confirmPassword}
+                    onChange={e => setConfirmPassword(e.target.value)}
                   />
                 </div>
